Refactor files to make searching easier
