+++ LuffyData/TaskData
@@ -1,6 +1,8 @@
-DEADLINE      | 0 | Submit Quiz | BY: Sep 02 2024 2359HRS
+DEADLINE      | 1 | Submit Quiz | BY: Sep 02 2024 2359HRS
 TO-DO         | 1 | read book
 TO-DO         | 1 | put book on shelf
 TO-DO         | 1 | book an appointment to doctor
 DEADLINE      | 0 | Do 2103T iP | BY: Sep 06 2024 2359HRS
 TO-DO         | 0 | CS2103T
+TO-DO         | 0 | homework
+TO-DO         | 0 | ldsa
