+++ src/main/java/luffy/Luffy.java
@@ -3,6 +3,15 @@ package luffy;
 import java.io.FileNotFoundException;
 import java.util.Scanner;
 
+import luffy.command.Command;
+import luffy.command.ExitCommand;
+import luffy.gui.LuffyLauncher;
+import luffy.parser.LuffyParser;
+import luffy.storage.Storage;
+import luffy.task.TaskList;
+import luffy.ui.LuffyUI;
+
+
 /**
  * Represents a chatbot that has various commands
  * for users to note down tasks and check them off
