+++ src/main/java/luffy/command/ListCommand.java
@@ -1,4 +1,8 @@
-package luffy;
+package luffy.command;
+
+import luffy.storage.Storage;
+import luffy.task.TaskList;
+import luffy.ui.LuffyUI;
 
 /**
  * Represents a command that displays the list of tasks
