+++ src/main/java/luffy/gui/LuffyMain.java
@@ -1,4 +1,4 @@
-package luffy;
+package luffy.gui;
 
 import java.io.IOException;
 
@@ -7,6 +7,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
+import luffy.Luffy;
 
 /**
  * A GUI for Duke using FXML.
