+++ src/main/java/luffy/storage/Storage.java
@@ -1,4 +1,5 @@
-package luffy;
+package luffy.storage;
+
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
@@ -7,8 +8,11 @@ import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 
-
-
+import luffy.task.Deadline;
+import luffy.task.Event;
+import luffy.task.Task;
+import luffy.task.TaskList;
+import luffy.task.Todo;
 
 /**
  * Represents a file storing system that
