+++ src/main/java/luffy/task/Deadline.java
@@ -1,4 +1,4 @@
-package luffy;
+package luffy.task;
 
 /**
  * Represents a task that has a deadline
