+++ src/main/java/luffy/task/TaskList.java
@@ -1,4 +1,4 @@
-package luffy;
+package luffy.task;
 
 import java.util.ArrayList;
